+++ privaxy/src/server/lib.rs
@@ -1,4 +1,5 @@
 use crate::blocker::AdblockRequester;
+use crate::configuration::NetworkConfig;
 use crate::proxy::exclusions::LocalExclusionStore;
 use crate::web_gui::events::Event;
 use hyper::server::conn::AddrStream;
@@ -9,11 +10,16 @@ use proxy::exclusions;
 use reqwest::redirect::Policy;
 use std::convert::Infallible;
 use std::env;
+use std::net::IpAddr;
 use std::net::SocketAddr;
+use std::str::FromStr;
 use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
+use tokio::signal::unix::{signal, SignalKind};
 use tokio::sync::broadcast;
+use tokio::sync::Notify;
+
 pub mod blocker;
 mod blocker_utils;
 mod ca;
@@ -22,10 +28,9 @@ pub mod configuration;
 mod proxy;
 pub mod statistics;
 mod web_gui;
-
 pub const WEBAPP_FRONTEND_DIR: Dir<'_> = include_dir!("web_frontend/dist");
 
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct PrivaxyServer {
     pub ca_certificate_pem: String,
     pub configuration_updater_sender: tokio::sync::mpsc::Sender<configuration::Configuration>,
@@ -37,16 +42,38 @@ pub struct PrivaxyServer {
     pub requests_broadcast_sender: broadcast::Sender<Event>,
 }
 
-// Helper function to parse the IP address string into an array of u8
-pub(crate) fn parse_ip_address(ip_str: &str) -> [u8; 4] {
-    let mut ip: [u8; 4] = [0, 0, 0, 0];
-    let parts: Vec<&str> = ip_str.split('.').collect();
-    for (i, part) in parts.iter().enumerate() {
-        if let Ok(num) = part.parse::<u8>() {
-            ip[i] = num;
+pub(crate) fn parse_ip_address(ip_str: &str) -> IpAddr {
+    IpAddr::from_str(ip_str).unwrap()
+}
+
+async fn handle_signals() -> (Arc<Notify>, Arc<Notify>) {
+    let notify_shutdown = Arc::new(Notify::new());
+    let notify_reload = Arc::new(Notify::new());
+    let notify_shutdown_clone = notify_shutdown.clone();
+    let notify_reload_clone = notify_reload.clone();
+
+    tokio::spawn(async move {
+        let mut hup_signal =
+            signal(SignalKind::hangup()).expect("failed to set up SIGHUP signal handler");
+        let mut term_signal =
+            signal(SignalKind::terminate()).expect("failed to set up SIGTERM signal handler");
+
+        loop {
+            tokio::select! {
+                _ = hup_signal.recv() => {
+                    log::info!("Received SIGHUP signal, restarting child processes...");
+                    notify_reload_clone.notify_waiters();
+                }
+                _ = term_signal.recv() => {
+                    log::info!("Received SIGTERM signal, shutting down gracefully...");
+                    notify_shutdown_clone.notify_waiters();
+                    std::process::exit(0);
+                }
+            }
         }
-    }
-    ip
+    });
+
+    (notify_shutdown, notify_reload)
 }
 
 pub async fn start_privaxy() -> PrivaxyServer {
@@ -129,58 +156,40 @@ pub async fn start_privaxy() -> PrivaxyServer {
     configuration_updater.start();
 
     let configuration_save_lock = Arc::new(tokio::sync::Mutex::new(()));
-    let ip = match env::var("PRIVAXY_IP_ADDRESS") {
-        Ok(val) =>
-        // Parse the IP address from the environment variable string
-        {
-            parse_ip_address(&val)
-        }
-        Err(_) =>
-        // Set a default IP address
-        {
-            parse_ip_address(&network_config.bind_addr.clone())
-        }
-    };
-    let web_api_server_addr = SocketAddr::from((ip, network_config.web_port));
 
-    let mut tls_cert = None;
-    let mut tls_key = None;
-    if network_config.tls {
-        tls_cert = match network_config
-            .read_or_create_tls_cert(ca_certificate.clone(), ca_private_key.clone())
-            .await
-        {
-            Ok(cert) => Some(cert),
-            Err(err) => {
-                panic!("Failed to read or create TLS certificate: {err}");
-            }
-        };
-        tls_key = match network_config.get_tls_key().await {
-            Ok(key) => Some(key),
-            Err(err) => {
-                panic!("Failed to read or create TLS key: {err}");
-            }
-        };
-    }
-    web_gui::start_frontend(
-        broadcast_tx.clone(),
-        statistics.clone(),
-        blocking_disabled_store.clone(),
-        configuration_updater_tx.clone(),
-        configuration_save_lock.clone(),
-        local_exclusion_store.clone(),
-        web_api_server_addr,
-        tls_cert,
-        tls_key,
-        network_config.tls,
-    );
+    let (notify_shutdown, notify_reload) = handle_signals().await;
 
+    let block_disable_ref = blocking_disabled_store.clone();
+    let local_exclusion_store_ref = local_exclusion_store.clone();
+    let stats_clone = statistics.clone();
+    let configuration_updater_tx_ref = configuration_updater_tx.clone();
+    let configuration_save_lock_ref = configuration_save_lock.clone();
+    let broadcast_tx_ref = broadcast_tx.clone();
+
+    tokio::spawn(async move {
+        loop {
+            let (_sig_tx, sig_rx) = tokio::sync::mpsc::channel::<tokio::signal::unix::Signal>(1);
+            log::info!("Starting Privaxy frontend");
+            privaxy_frontend(
+                broadcast_tx_ref.clone(),
+                local_exclusion_store_ref.clone(),
+                stats_clone.clone(),
+                block_disable_ref.clone(),
+                configuration_updater_tx_ref.clone(),
+                configuration_save_lock_ref.clone(),
+                sig_rx,
+                notify_reload.clone(),
+            )
+            .await;
+            notify_reload.notified().await;
+            log::info!("Stopping Privaxy frontend");
+        }
+    });
+
+    let disabled_store_ref = blocking_disabled_store_clone.clone();
     thread::spawn(move || {
-        let blocker = blocker::Blocker::new(
-            crossbeam_sender,
-            crossbeam_receiver,
-            blocking_disabled_store,
-        );
+        let blocker =
+            blocker::Blocker::new(crossbeam_sender, crossbeam_receiver, disabled_store_ref);
 
         blocker.handle_requests()
     });
@@ -225,6 +234,7 @@ pub async fn start_privaxy() -> PrivaxyServer {
         }
     });
 
+    let ip = env_or_config_ip(&network_config).await;
     let proxy_server_addr = SocketAddr::from((ip, network_config.proxy_port));
 
     let server = Server::bind(&proxy_server_addr)
@@ -234,9 +244,6 @@ pub async fn start_privaxy() -> PrivaxyServer {
         .serve(make_service);
 
     log::info!("Proxy available at http://{}", proxy_server_addr);
-    log::info!("Web server available at http://{}", web_api_server_addr);
-    log::info!("API server available at http://{}/api", web_api_server_addr);
-
     if let Err(e) = server.await {
         log::error!("server error: {}", e);
     }
@@ -244,10 +251,97 @@ pub async fn start_privaxy() -> PrivaxyServer {
     PrivaxyServer {
         ca_certificate_pem,
         configuration_updater_sender: configuration_updater_tx,
-        configuration_save_lock: configuration_save_lock,
+        configuration_save_lock,
         blocking_disabled_store: blocking_disabled_store_clone,
         statistics: statistics_clone,
         local_exclusion_store: local_exclusion_store_clone,
         requests_broadcast_sender: broadcast_tx_clone,
+    }
+}
+
+async fn privaxy_frontend(
+    broadcast_tx: tokio::sync::broadcast::Sender<Event>,
+    local_exclusion_store: LocalExclusionStore,
+    statistics: statistics::Statistics,
+    block_disable_ref: blocker::BlockingDisabledStore,
+    configuration_updater_tx: tokio::sync::mpsc::Sender<configuration::Configuration>,
+    configuration_save_lock: Arc<tokio::sync::Mutex<()>>,
+    mut sig_rx: tokio::sync::mpsc::Receiver<tokio::signal::unix::Signal>,
+    notify_reload: Arc<tokio::sync::Notify>,
+) {
+    let frontend = web_gui::get_frontend(
+        broadcast_tx.clone(),
+        statistics.clone(),
+        &block_disable_ref,
+        &configuration_updater_tx,
+        &configuration_save_lock,
+        &local_exclusion_store,
+        notify_reload.clone(),
+    );
+    let frontend_server = warp::serve(frontend);
+    let config = read_configuration(&configuration_save_lock).await;
+    let ip = env_or_config_ip(&config.network).await;
+    let web_api_server_addr = SocketAddr::from((ip, config.network.web_port));
+    if config.network.tls {
+        let lock = configuration_save_lock.lock().await;
+        let ca_certificate = config.ca.get_ca_certificate().await.unwrap();
+        let ca_private_key = config.ca.get_ca_private_key().await.unwrap();
+        drop(lock);
+        let tls_cert = match config
+            .network
+            .read_or_create_tls_cert(ca_certificate.clone(), ca_private_key.clone())
+            .await
+        {
+            Ok(cert) => cert,
+            Err(err) => {
+                panic!("Failed to read or create TLS certificate: {err}");
+            }
+        };
+        let tls_key = match config.network.get_tls_key().await {
+            Ok(key) => key,
+            Err(err) => {
+                panic!("Failed to read or create TLS key: {err}");
+            }
+        };
+        tokio::spawn(async move {
+            let (_, task) = frontend_server
+                .tls()
+                .cert(tls_cert.to_pem().unwrap())
+                .key(tls_key.private_key_to_pem_pkcs8().unwrap())
+                .bind_with_graceful_shutdown(web_api_server_addr, async move {
+                    notify_reload.clone().notified().await;
+                });
+            log::info!("Web server available at https://{web_api_server_addr}/");
+            log::info!("API server available at https://{web_api_server_addr}/api");
+
+            task.await;
+        });
+    } else {
+        tokio::spawn(async move {
+            let (_, task) =
+                frontend_server.bind_with_graceful_shutdown(web_api_server_addr, async move {
+                    let _ = sig_rx.recv().await;
+                });
+            log::info!("Web server available at http://{web_api_server_addr}/");
+            log::info!("API server available at http://{web_api_server_addr}/api");
+            task.await
+        });
+    }
+}
+
+async fn read_configuration(
+    configuration_save_lock: &Arc<tokio::sync::Mutex<()>>,
+) -> configuration::Configuration {
+    let lock = configuration_save_lock.lock().await;
+    let config = configuration::Configuration::read_from_home()
+        .await
+        .unwrap();
+    drop(lock);
+    config
+}
+async fn env_or_config_ip(network_config: &NetworkConfig) -> IpAddr {
+    match env::var("PRIVAXY_IP_ADDRESS") {
+        Ok(val) => parse_ip_address(&val),
+        Err(_) => network_config.parsed_ip_address(),
     }
 }