+++ privaxy/src/server/web_gui/settings/network.rs
@@ -1,7 +1,11 @@
 use super::get_error_response;
+use crate::configuration;
 use crate::configuration::NetworkConfig;
 use crate::web_gui::with_configuration_save_lock;
 use crate::web_gui::with_configuration_updater_sender;
+use serde::{Deserialize, Serialize};
+use tokio::sync::Notify;
+
 use std::convert::Infallible;
 use std::sync::Arc;
 use tokio::sync::mpsc::Sender;
@@ -10,6 +14,33 @@ use warp::http::Response;
 use warp::Filter as RouteFilter;
 
 use crate::configuration::Configuration;
+
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
+/// Network configuration for Privaxy
+pub struct NetworkConfigRequest {
+    /// Bind address for the proxy server.
+    pub bind_addr: String,
+    /// Port for the proxy server.
+    pub proxy_port: u16,
+    /// Port for the web server.
+    pub web_port: u16,
+    /// Enable TLS for the web server.
+    pub tls: bool,
+}
+
+impl Into<NetworkConfig> for NetworkConfigRequest {
+    fn into(self) -> NetworkConfig {
+        NetworkConfig {
+            bind_addr: self.bind_addr,
+            proxy_port: self.proxy_port,
+            web_port: self.web_port,
+            tls: self.tls,
+            tls_cert_path: None,
+            tls_key_path: None,
+            listen_url: None,
+        }
+    }
+}
 
 async fn get_network_settings() -> Result<Box<dyn warp::Reply>, Infallible> {
     log::debug!("Getting network settings");
@@ -24,27 +55,42 @@ async fn get_network_settings() -> Result<Box<dyn warp::Reply>, Infallible> {
 }
 
 async fn put_network_settings(
-    network_settings: NetworkConfig,
+    network_settings: NetworkConfigRequest,
     configuration_updater_sender: Sender<Configuration>,
     configuration_save_lock: Arc<tokio::sync::Mutex<()>>,
+    notify_reload: Arc<Notify>,
 ) -> Result<Box<dyn warp::Reply>, Infallible> {
-    let _guard = configuration_save_lock.lock().await;
+    let lock = configuration_save_lock.lock().await;
     let mut configuration = match Configuration::read_from_home().await {
         Ok(configuration) => configuration,
         Err(err) => {
-            log::error!("Failed to get network settings: {err}");
+            log::error!("Failed to get network settings: {}", err);
             return Ok(Box::new(get_error_response(err)));
         }
     };
-    if let Err(err) = configuration.set_network_settings(&network_settings).await {
-        log::error!("Invalid network settings: {err}");
+
+    drop(lock);
+    let mut net_cfg: NetworkConfig = network_settings.into();
+    let current_cfg = configuration.network.clone();
+    net_cfg.tls_cert_path = current_cfg.tls_cert_path;
+    net_cfg.tls_key_path = current_cfg.tls_key_path;
+    net_cfg.listen_url = current_cfg.listen_url;
+    if let Err(err) = &net_cfg.validate().await {
+        log::error!("Invalid network settings: {}", err);
         return Ok(Box::new(get_error_response(err)));
     };
+    configuration.network = net_cfg.clone();
 
+    let guard = configuration_save_lock.lock().await;
+    configuration.save().await.unwrap();
     configuration_updater_sender
         .send(configuration.clone())
         .await
         .unwrap();
+    drop(guard);
+
+    notify_reload.notify_waiters(); // Notify the reload signal
+
     Ok(Box::new(
         Response::builder()
             .status(http::StatusCode::NO_CONTENT)
@@ -55,15 +101,35 @@ async fn put_network_settings(
 pub(super) fn create_routes(
     configuration_updater_sender: Sender<Configuration>,
     configuration_save_lock: Arc<tokio::sync::Mutex<()>>,
+    notify_reload: Arc<Notify>,
 ) -> BoxedFilter<(impl warp::Reply,)> {
-    warp::path::end()
-        .and(warp::get().and_then(self::get_network_settings))
-        .or(warp::put()
-            .and(warp::body::json())
-            .and(with_configuration_updater_sender(
-                configuration_updater_sender,
-            ))
-            .and(with_configuration_save_lock(configuration_save_lock))
-            .and_then(self::put_network_settings))
-        .boxed()
+    let get_route = warp::get()
+        .and(warp::path::end())
+        .and_then(get_network_settings);
+
+    let put_route = warp::put()
+        .and(warp::path::end())
+        .and(warp::body::json())
+        .and(with_configuration_updater_sender(
+            configuration_updater_sender.clone(),
+        ))
+        .and(with_configuration_save_lock(
+            configuration_save_lock.clone(),
+        ))
+        .and_then(
+            move |network_settings, configuration_updater_sender, configuration_save_lock| {
+                let notify_reload = notify_reload.clone();
+                async move {
+                    put_network_settings(
+                        network_settings,
+                        configuration_updater_sender,
+                        configuration_save_lock,
+                        notify_reload,
+                    )
+                    .await
+                }
+            },
+        );
+
+    get_route.or(put_route).boxed()
 }