+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 - Static files and API are now under the same route
   - Static files are found as before
   - API calls are under `/api`
+- Privaxy now honors SIGHUP
+  - `systemctl reload privaxy` will reload privaxys configuration
+- Change settings in webserver
+  - Upon a successful change, the server will reload
 
 
 ## v0.5.2